clean up config
