+++ employee/src/main/java/com/emp/config/WebSecurityConfig.java
@@ -1,27 +1,20 @@
 package com.emp.config;
 
 
-import javax.servlet.Filter;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.filter.ShallowEtagHeaderFilter;
-@Configuration
+
 @EnableWebSecurity//enable Spring Security’s web security support and provide the Spring MVC integration
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 	
+
 	@Autowired
-    private UserDetailsService userDetailsService;
+	private DaoAuthenticationProvider authProvider;
 	
 	/**
 	 * The configure(HttpSecurity) method defines which URL paths should be secured
@@ -52,36 +45,10 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     
     @Override
     protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.authenticationProvider(authProvider());
+        auth.authenticationProvider(authProvider);
     }
   
     
-  @Bean
-  public DaoAuthenticationProvider authProvider() {
-      DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-      authProvider.setUserDetailsService(userDetailsService);
-      /*basically I am specifying that the BCryptPasswordEncoder(BCrypt algorithm) 
-       * must be used to authenticate the user against the hashed password in the DB.
-       * That is the user who will log in with plain text password will be hashed using 
-       * this algorithm .This hashed password will then be compared against the one 
-       * stored in the password column of user table.
-       */
-      authProvider.setPasswordEncoder(passwordEncoder());
-      return authProvider;
-  }
     
-    @Bean
-   	public PasswordEncoder passwordEncoder(){
-   		PasswordEncoder encoder = new BCryptPasswordEncoder();
-   		return encoder;
-   	}
-    
-    /**
-     * add etag header element for GET request.Can be used for reducing network bandwidth and for concurrency
-     * @return
-     */
-    @Bean
-    public Filter shallowEtagHeaderFilter() {
-        return new ShallowEtagHeaderFilter();
-    }
+   
 }
